requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
dnspython==2.2.0
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.31
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
